Adding some files/practices for beginners
Higher order functions part is great for all the beginners to practice. If you still feel unsure about your knowledge, repeat them. There is a little test, it is solved, but it's a good starting point for creating your own tests. Then there is OOP and Async/Await. More to come.